Fixed header icon spacing on android and renamed home screen and keyboard for clarity
