+++ lib/app_core/theme/sizes.dart
@@ -6,7 +6,7 @@ class Paddings {
 
   static const listGap = 16.0;
 
-  static const scaffoldTop = 96.0;
+  static const scaffoldTop = 4.0;
   static const scaffold = 24.0;
 }
