+++ lib/use_cases/home/ui/compare_currency_card.dart
@@ -4,7 +4,7 @@ import 'package:travelconverter/app_core/widgets/currency_card.dart';
 import 'package:travelconverter/app_core/widgets/gap.dart';
 import 'package:travelconverter/l10n/app_localizations.dart';
 import 'package:travelconverter/services/logger.dart';
-import 'package:travelconverter/use_cases/main_screen/ui/compare_keyboard_sheet.dart';
+import 'package:travelconverter/use_cases/home/ui/custom_keyboard_sheet.dart';
 import 'package:travelconverter/widgets/animate_in.dart';
 import 'package:travelconverter/model/currency.dart';
 import 'package:travelconverter/state_container.dart';
@@ -99,7 +99,7 @@ class _CompareCurrencyCardState extends State<CompareCurrencyCard>
     double? value = await showModalBottomSheet(
         context: context,
         isScrollControlled: true,
-        builder: (ctx) => CompareKeyboardSheet(
+        builder: (ctx) => CustomKeyboardSheet(
             currencyCode: widget.currency.code, initialValue: currentValue));
 
     if (value != null) {
