+++ test/currency_convert_card_widget_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:travelconverter/use_cases/main_screen/ui/compare_currency_card.dart';
+import 'package:travelconverter/use_cases/home/ui/compare_currency_card.dart';
 
 import 'mocks/mock_app_container_builder.dart';
 import 'mocks/mock_currency.dart';
